Remove unneeded ASN.1 code, fix get_cert_info/1
get_cert_info/1 fails for newer types of push certs, such as Apple Watch
and Web Push Certs.

This commit fixes that issue, and now works for the following Issuer CN
types:

- Apple Development IOS Push Services
- Apple Development Push Services
- Apple Production IOS Push Services
- Apple Production Push Services
- Apple Push Services
- Pass Type ID
- VoIP Services
- WatchKit Services
- Website Push ID

In addition, the ASN.1 parsing was giving intractable Dialyzer errors,
and much of it was replaced by calling asn1rt_nif:decode_ber_tlv/1,
which is undocumented, but the asn1rt module is deprecated, so this will
have to do in the meantime.