Fix signing bug in apns_jwt
Signed JWTs are not accepted by APNS. On investigation, it turns
out that the way the private key is parsed does not go far enough.
What was thought to be the private key was, in fact, an octet string
that was itself ASN.1 as defined in RFC 5915.

This octet string needs to be parsed into an ECPrivateKey structure,
and that structure used for signing.

This commit fixes that issue. The generated JWTs are accepted by APNS
now.

This commit also improves the way the keys are generated. Instead of
always needing to parse the private key PEM data every time, the parsed
information is now kept in a context object, which the caller can
maintain indefinitely and pass to jwt/1 as needed to generate a new JWT.